Revert "Front-End Update Admin Page"
This reverts commit 99855075b0a2386aaebb672a927f967e7a1cd69b.

+++ src/main/java/com/logistics/logisticsCompany/service/EmployeeServiceImpl.java
@@ -5,16 +5,13 @@ import com.logistics.logisticsCompany.DTO.EntityDtoMapper;
 import com.logistics.logisticsCompany.customExceptions.EntityNotFoundException;
 import com.logistics.logisticsCompany.entities.logisticsCompany.LogisticsCompany;
 import com.logistics.logisticsCompany.entities.offices.Office;
-import com.logistics.logisticsCompany.entities.orders.Shipment;
 import com.logistics.logisticsCompany.entities.users.Employee;
 import com.logistics.logisticsCompany.entities.users.User;
 import com.logistics.logisticsCompany.repository.EmployeeRepository;
 import com.logistics.logisticsCompany.repository.LogisticsCompanyRepository;
 import com.logistics.logisticsCompany.repository.OfficeRepository;
-import com.logistics.logisticsCompany.repository.ShipmentRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
 import java.util.Optional;
@@ -43,25 +40,21 @@ public class EmployeeServiceImpl implements EmployeeService {
     
     private final UserLinkageService userLinkageService;
     
-    private final ShipmentRepository shipmentRepository;
-    
     @Autowired
-    public EmployeeServiceImpl(EmployeeRepository employeeRepository, LogisticsCompanyRepository logisticsCompanyRepository, UserLinkageService userLinkageService, OfficeRepository officeRepository, ShipmentRepository shipmentRepository){
+    public EmployeeServiceImpl(EmployeeRepository employeeRepository, LogisticsCompanyRepository logisticsCompanyRepository, UserLinkageService userLinkageService, OfficeRepository officeRepository){
         this.employeeRepository = employeeRepository;
         this.logisticsCompanyRepository = logisticsCompanyRepository;
         this.userLinkageService = userLinkageService;
         this.officeRepository = officeRepository;
-        this.shipmentRepository = shipmentRepository;
     }
  
     /**
      * Creates a new employee.
-     *
      * @param employeeDTO the employee to create
      */
     
     @Override
-    public void createEmployee(EmployeeDTO employeeDTO) {
+    public Employee createEmployee(EmployeeDTO employeeDTO) {
         Employee employee = EntityDtoMapper.convertEmployeeDtoToEntity(employeeDTO);
 
         
@@ -86,7 +79,7 @@ public class EmployeeServiceImpl implements EmployeeService {
             employee.setCurrentOffice(office);
         }
         
-        employeeRepository.save(employee);
+        return employeeRepository.save(employee);
     }
 
     /**
@@ -123,28 +116,14 @@ public class EmployeeServiceImpl implements EmployeeService {
      * Deletes an existing employee.
      * @param employeeId the id of the employee to delete
      */
-    @Transactional
     @Override
     public void deleteEmployee(long employeeId) {
-        // Fetch shipments where the employee is the receiver
-        List<Shipment> receivedShipments = shipmentRepository.findByReceiverEmployeeId(employeeId);
-        receivedShipments.forEach(shipment -> {
-            shipment.setReceiverEmployee(null);
-            shipmentRepository.save(shipment);
-        });
-        
-        // Fetch shipments where the employee is the sender
-        List<Shipment> sentShipments = shipmentRepository.findBySenderEmployeeId(employeeId);
-        sentShipments.forEach(shipment -> {
-            shipment.setSenderEmployee(null);
-            shipmentRepository.save(shipment);
-        });
-        
-        // After dissociating the employee from all shipments, delete the employee
+        if (!employeeRepository.existsById(employeeId)) {
+            throw new EntityNotFoundException("Employee with the provided id doesn't exist");
+        }
         employeeRepository.deleteById(employeeId);
     }
-    
-    
+
     /**
      * Assigns an office to an employee.
      * @param employeeId the id of the employee
